requirements.txt
pyTelegramBotAPI
datetime
pydub
python-decouple
gTTS
